Safari-611.1.10/LayoutTests/fast/css/padding-calc-value.html

<!DOCTYPE html>
<body>
<script src="../../resources/js-test-pre.js"></script>

<div id="testDiv" style="padding: 0 calc(.5em - 2px) 0 0; font-size: 26px;">
</div>

<script>
description("Tests assigning a calculated value to 'padding' CSS property.");

var testDiv = document.getElementById("testDiv");
shouldBeEqualToString("window.getComputedStyle(testDiv).getPropertyValue('padding-right')", "11px");

</script>
<script src="../../resources/js-test-post.js"></script>
</body>
