Safari-611.1.10/LayoutTests/fast/css/padding-calc-value-expected.txt

Tests assigning a calculated value to 'padding' CSS property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.getComputedStyle(testDiv).getPropertyValue('padding-right') is "11px"
PASS successfullyParsed is true

TEST COMPLETE
